Move lend filter experiment handling to mixin and apply to the lend/filter and /lend-by-cat/##loan-channel## page. This allows activation of the lend filter experiment when arriving by redirect from /lend/##loan-channel## pages.
